In the name of God.
The repo contains two npm projects. The first is the client-side module that runs in the browser and can be used via `<script>` tag or as an ES6 module. It helps the user of web application to edit some special parts of the page (`contenteditable`). It's a Webpack + Typescript project. Its basic idea is published on MDN website:
https://developer.mozilla.org/en-US/docs/Web/Guide/HTML/Editable_content#Example_A_simple_but_complete_rich_text_editor

The second is the server-side module that runs on the Node.js platform. It must be run on the same machine that source files of the web application located. It serves the client by saving edited contents into suitable places of the source files. It's a Typescript project and powered by `express` and `cors` as the server, `jsdom`, `yargs`, `source-map-support`, etc.